move tests to test_arrow.py
